Description:
In this project, I was asked to calculate the amount of time it took for the user to complete typing out a random sentence. I used the random.randint function to print out three random sentences for the users. Once they had finished typing, the code would tell the user their time in seconds.