Description:
In this project, I was asked to code a game where users make as many words as they can with the seven random letters (somewhat like Scrabble). Each word that is made must be a valid word, or else no points will be credited. The user gains points based on how many letters there are in the valid word.