Description:
In this project, I was asked to code a game of professional Blackjack vs the dealer. In this project, I was the player, and the computer was playing Blackjack as the dealer. I used the "H" to represent "Hit" and "S" to represent "Stay". Each competitor had started with two cards, and from there they got to decide if they wanted to hit or stay, as the objective of the game was to see who got closest to the number, twenty-one. After comparing the two scores, the code would print out a statement announcing who had won and lost.