Description
In this project, I was asked to use recursions to check if a word was a palindrome or not. A palindrome word is a word that is spelled out the same forwards and backward, so a great example would be "racecar". The recursion process used sublists to check the word from the right side and left side, and printed out either "True" or "False".